feat : 인증, 권한 설정

+++ drugsafe_backend/medicines/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, get_object_or_404
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from .models import Drug, Review, Score, Comment
 from .serializers import DrugListSerializer, CommentSerializer, ReviewSerializer, ReviewDetailSerializer, ScoreSerializer
@@ -26,6 +27,7 @@ def drug_list(request):
 #             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 @api_view(['GET','POST'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def review_list(request, drug_pk):
     drug = get_object_or_404(Drug, pk=drug_pk)
 
@@ -42,6 +44,7 @@ def review_list(request, drug_pk):
 
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def review_detail(request, drug_pk, review_pk):
     review = get_object_or_404(
         Review,
@@ -52,16 +55,21 @@ def review_detail(request, drug_pk, review_pk):
         serializer = ReviewDetailSerializer(review)
         return Response(serializer.data)
     elif request.method == 'PUT':
+        if review.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         serializer = ReviewSerializer(review, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
     elif request.method == 'DELETE':
+        if review.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         review.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['GET', 'POST'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def comment_list(request, review_pk):
     review = get_object_or_404(Review, pk=review_pk)
     if request.method == 'GET':
@@ -81,6 +89,7 @@ def comment_list(request, review_pk):
 
 
 @api_view(['PUT', 'DELETE'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def comment_detail(request, review_pk, comment_pk):
     comment = get_object_or_404(
         Comment,
@@ -88,17 +97,22 @@ def comment_detail(request, review_pk, comment_pk):
         review_id=review_pk
     )
     if request.method == 'PUT':
+        if comment.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         serializer = CommentSerializer(comment, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
 
     elif request.method == 'DELETE':
+        if comment.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['GET', 'POST'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def score_list(request, drug_pk):
     drug = get_object_or_404(Drug, pk=drug_pk)
     if request.method == 'GET':
@@ -117,6 +131,7 @@ def score_list(request, drug_pk):
 
 
 @api_view(['PUT', 'DELETE'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def score_detail(request, drug_pk, score_pk):
     score = get_object_or_404(
         Score,
@@ -124,11 +139,15 @@ def score_detail(request, drug_pk, score_pk):
         drug_id=drug_pk
     )
     if request.method == 'PUT':
+        if score.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         serializer = ScoreSerializer(score, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
 
     elif request.method == 'DELETE':
+        if score.user != request.user:
+            return Response({'error': '권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
         score.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
